pyproject.toml
[project]
name = "job-tracker-mcp"
version = "0.1.0"
description = "Job application tracker MCP server (FastMCP + Prefab UI)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp[apps]>=2.0.0",
    "prefab-ui>=0.19.0,<0.20",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "filelock>=3.16.0",
    "python-dotenv>=1.0.0",
]
[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
streamlit-client = [
    "streamlit>=1.38.0",
    "mcp>=1.2.0",
    "openai>=1.54.0",
]
web = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "mcp>=1.2.0",
    "openai>=1.54.0",
]

[project.scripts]
job-tracker-mcp = "job_tracker_mcp.server:main"

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
